bin/requirements.txt
Click==7.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
psycopg2==2.8.4
pynmcli==1.0.5
pyparsing==2.4.5
python-crontab==2.4.0
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.4.0
six==1.13.0
Werkzeug==0.16.0
